// bank-server/src/interfaces/operations/game-operation-type.enum.ts
export enum GameOperationType {
    SEND_MONEY_TO_PLAYER = 'SEND_MONEY_TO_PLAYER',
    SEND_MONEY_TO_BANK = 'SEND_MONEY_TO_BANK',
    GET_MONEY_FROM_BANK = 'GET_MONEY_FROM_BANK',
    TOTAL_WRITE_OFF = 'TOTAL_WRITE_OFF',
    TOTAL_PAYMENT = 'TOTAL_PAYMENT',
    GET_CREDIT = 'GET_CREDIT',
    REPAYMENT_CREDIT = 'REPAYMENT_CREDIT',
}